refactor dapp naming - [CORE-824]
