+++ dapps/assets/src/components/Assets.vue
@@ -43,9 +43,9 @@
           </div>
         </evan-dapp-wrapper-level-2>
         <transition name="fade" mode="out-in">
-          <div class="container content">
+          <!-- <div class="container content"> -->
             <router-view></router-view>
-          </div>
+          <!-- </div> -->
         </transition>
       </template>
     </evan-dapp-wrapper>
