+++ dapps/digital-twin-detail/src/components/DigitalTwinDetail.scss
