+++ dapps/digital-twin-detail/src/components/DigitalTwinDetail.ts
@@ -24,7 +24,7 @@ import Component, { mixins } from 'vue-class-component';
 import { EvanComponent } from '@evan.network/ui-vue-core';
 
 @Component
-export default class EvanTwinDetailComponent extends mixins(EvanComponent) {
+export default class DigitalTwinDetailComponent extends mixins(EvanComponent) {
   mounted() {
 
   }
