+++ dapps/digital-twin-detail/src/components/data/DetailDataGeneral.vue
