+++ dapps/digital-twin-detail/src/i18n/translations.ts
@@ -17,8 +17,8 @@
   the following URL: https://evan.network/license/
 */
 
-import de from './de.json';
-import en from './en.json';
+import de from './de.json.js';
+import en from './en.json.js';
 
 // map all languages
 export default { de, en };
