+++ dapps/digital-twin-detail/tsconfig.json
