+++ dapps/digital-twin-detail/webpack.config.js
@@ -18,7 +18,7 @@
 */
 
 module.exports = require('../../scripts/dapp/webpack.config')(
-  require('./dbcp.json').public.name,
+  require('./dbcp.json.js').public.name,
   require('path').resolve(__dirname, './dist'),
   true,
   false,
